feat: add "Buffer" as supported read option
