+++ nodejs-polars/src/dataframe/io.rs
@@ -1,5 +1,5 @@
 use crate::conversion::prelude::*;
-use crate::dataframe::finish_from_rows;
+use polars::frame::row::{rows_to_schema, Row};
 use crate::error::JsPolarsEr;
 use crate::file::JsWriteStream;
 use crate::prelude::JsResult;
@@ -7,10 +7,10 @@ use napi::{
     CallContext, JsExternal, JsObject, JsString, JsUndefined,
     JsUnknown,
 };
-use polars::frame::row::Row;
 use polars::prelude::*;
 use std::fs::File;
 use std::io::{BufReader, Cursor};
+use std::path::{Path, PathBuf};
 
 
 #[js_function(1)]
@@ -409,4 +409,30 @@ pub(crate) fn read_array_rows(cx: CallContext) -> JsResult<JsExternal> {
         })
         .collect();
     finish_from_rows(rows)?.try_into_js(&cx)
+}
+
+
+fn resolve_homedir(path: &Path) -> PathBuf {
+    // replace "~" with home directory
+    if path.starts_with("~") {
+        if let Some(homedir) = dirs::home_dir() {
+            return homedir.join(path.strip_prefix("~").unwrap());
+        }
+    }
+
+    path.into()
+}
+fn finish_from_rows(rows: Vec<Row>) -> JsResult<DataFrame> {
+    let schema = rows_to_schema(&rows);
+    let fields = schema
+        .fields()
+        .iter()
+        .map(|fld| match fld.data_type() {
+            DataType::Null => Field::new(fld.name(), DataType::Boolean),
+            _ => fld.clone(),
+        })
+        .collect();
+    let schema = Schema::new(fields);
+
+    DataFrame::from_rows_and_schema(&rows, &schema).map_err(|err| JsPolarsEr::from(err).into())
 }
