+++ nodejs-polars/src/dataframe/mod.rs
@@ -76,16 +76,18 @@ impl JsDataFrame {
             napi::Property::new(env, "tail")?.with_method(df::tail),
             napi::Property::new(env, "take_with_series")?.with_method(df::take_with_series),
             napi::Property::new(env, "take")?.with_method(df::take),
-            napi::Property::new(env, "to_js")?.with_method(io::to_js),
-            napi::Property::new(env, "to_json")?.with_method(io::to_json),
-            napi::Property::new(env, "to_row")?.with_method(io::to_row),
-            napi::Property::new(env, "to_rows")?.with_method(io::to_rows),
+
             napi::Property::new(env, "transpose")?.with_method(df::transpose),
             napi::Property::new(env, "var")?.with_method(df::var),
             napi::Property::new(env, "vstack")?.with_method(df::vstack),
             napi::Property::new(env, "width")?.with_method(df::width),
             napi::Property::new(env, "with_column")?.with_method(df::with_column),
             napi::Property::new(env, "with_row_count")?.with_method(df::with_row_count),
+            
+            napi::Property::new(env, "to_js")?.with_method(io::to_js),
+            napi::Property::new(env, "to_json")?.with_method(io::to_json),
+            napi::Property::new(env, "to_row")?.with_method(io::to_row),
+            napi::Property::new(env, "to_rows")?.with_method(io::to_rows),
             napi::Property::new(env, "write_csv_path")?.with_method(io::write_csv_path),
             napi::Property::new(env, "write_csv_stream")?.with_method(io::write_csv_stream),
             napi::Property::new(env, "write_json_path")?.with_method(io::write_json_path),
